Move home components into feature dir
