+++ app/page.tsx
@@ -1,5 +1,1 @@
-import HomePage from './components/HomePage';
-
-export default function Page() {
-  return <HomePage />;
-}
+export { default } from './features/home/page';
